==> 02_menu_data_generation.ipynb <==
# Menu Dataset for 'A Pizza restaurant

import pandas as pd
from faker import Faker

fake = Faker()

menu_items = [
    {"Item_ID": 1, "Item_Name": "Margherita", "Category": "Pizza", "Price": 32, "Available": True},
    {"Item_ID": 2, "Item_Name": "Diavola", "Category": "Pizza", "Price": 35, "Available": True},
    {"Item_ID": 3, "Item_Name": "Vegana", "Category": "Pizza", "Price": 36, "Available": True},
    {"Item_ID": 4, "Item_Name": "Amatriciana", "Category": "Pasta", "Price": 37, "Available": True},
    {"Item_ID": 5, "Item_Name": "Carbonara", "Category": "Pasta", "Price": 38, "Available": True},
    {"Item_ID": 6, "Item_Name": "Bolognese", "Category": "Pasta", "Price": 39, "Available": True},
    {"Item_ID": 7, "Item_Name": "Focaccia", "Category": "Appetizer", "Price": 15, "Available": True},
    {"Item_ID": 8, "Item_Name": "Coca-Cola", "Category": "Beverage", "Price": 10, "Available": True},
    {"Item_ID": 9, "Item_Name": "Tiramisu", "Category": "Dessert", "Price": 18, "Available": True},
    {"Item_ID": 10, "Item_Name": "Espresso", "Category": "Beverage", "Price": 12, "Available": True},
    {"Item_ID": 11, "Item_Name": "Water", "Category": "Beverage", "Price": 5, "Available": True}

]

menu_data = pd.DataFrame(menu_items)

menu_data.to_csv("menu_data.csv", index=False)

menu_data
