==> 03_inventory_data.ipynb <==
# Inventory ingredients Dataset for 'A Pizza restaurant

import pandas as pd

inventory_ingredients = [
    {"Item_ID": 1, "Ingredient": "Tomato Sauce", "Quantity_Per_Item": 100, "Stock_Level": 2000},  # Margherita
    {"Item_ID": 1, "Ingredient": "Cheese", "Quantity_Per_Item": 150, "Stock_Level": 1500},  # Margherita
    {"Item_ID": 1, "Ingredient": "Basil", "Quantity_Per_Item": 10, "Stock_Level": 500},  # Margherita

    {"Item_ID": 2, "Ingredient": "Tomato Sauce", "Quantity_Per_Item": 100, "Stock_Level": 2000},  # Diavola
    {"Item_ID": 2, "Ingredient": "Cheese", "Quantity_Per_Item": 150, "Stock_Level": 1600},  # Diavola
    {"Item_ID": 2, "Ingredient": "Salami", "Quantity_Per_Item": 50, "Stock_Level": 600},  # Diavola

    {"Item_ID": 3, "Ingredient": "Tomato Sauce", "Quantity_Per_Item": 100, "Stock_Level": 1800},  # Vegana
    {"Item_ID": 3, "Ingredient": "Vegan Cheese", "Quantity_Per_Item": 150, "Stock_Level": 1400},  # Vegana
    {"Item_ID": 3, "Ingredient": "Mushrooms", "Quantity_Per_Item": 50, "Stock_Level": 500},  # Vegana
    {"Item_ID": 3, "Ingredient": "Bell Peppers", "Quantity_Per_Item": 30, "Stock_Level": 400},  # Vegana

    {"Item_ID": 4, "Ingredient": "Pasta", "Quantity_Per_Item": 100, "Stock_Level": 1000},  # Amatriciana
    {"Item_ID": 4, "Ingredient": "Pancetta", "Quantity_Per_Item": 80, "Stock_Level": 800},  # Amatriciana
    {"Item_ID": 4, "Ingredient": "Tomato Sauce", "Quantity_Per_Item": 100, "Stock_Level": 1500},  # Amatriciana

    {"Item_ID": 5, "Ingredient": "Pasta", "Quantity_Per_Item": 100, "Stock_Level": 1100},  # Carbonara
    {"Item_ID": 5, "Ingredient": "Pancetta", "Quantity_Per_Item": 80, "Stock_Level": 850},  # Carbonara
    {"Item_ID": 5, "Ingredient": "Eggs", "Quantity_Per_Item": 2, "Stock_Level": 400},  # Carbonara

    {"Item_ID": 6, "Ingredient": "Pasta", "Quantity_Per_Item": 100, "Stock_Level": 900},  # Bolognese
    {"Item_ID": 6, "Ingredient": "Ground Beef", "Quantity_Per_Item": 120, "Stock_Level": 600},  # Bolognese
    {"Item_ID": 6, "Ingredient": "Tomato Sauce", "Quantity_Per_Item": 100, "Stock_Level": 1300},  # Bolognese

    {"Item_ID": 7, "Ingredient": "Flour", "Quantity_Per_Item": 50, "Stock_Level": 600},  # Focaccia
    {"Item_ID": 7, "Ingredient": "Olive Oil", "Quantity_Per_Item": 30, "Stock_Level": 400},  # Focaccia
    {"Item_ID": 7, "Ingredient": "Rosemary", "Quantity_Per_Item": 5, "Stock_Level": 200},  # Focaccia

    {"Item_ID": 8, "Ingredient": "Coca-Cola", "Quantity_Per_Item": 1, "Stock_Level": 500},  # Coca-Cola
    {"Item_ID": 9, "Ingredient": "Mascarpone", "Quantity_Per_Item": 100, "Stock_Level": 200},  # Tiramisu
    {"Item_ID": 9, "Ingredient": "Coffee", "Quantity_Per_Item": 50, "Stock_Level": 300},  # Tiramisu
    {"Item_ID": 10, "Ingredient": "Coffee", "Quantity_Per_Item": 30, "Stock_Level": 600},  # Espresso
    {"Item_ID": 11, "Ingredient": "Water", "Quantity_Per_Item": 1, "Stock_Level": 1000}  # Water
]

inventory_ingredients_df = pd.DataFrame(inventory_ingredients)

inventory_ingredients_df.to_csv("inventory_ingredients.csv", index=False)

inventory_ingredients_df
