==> 04_reviews_data.ipynb <==
# Reviews Dataset for 'A Pizza restaurant

!pip install requests

import requests
import pandas as pd

api_key = 'AIzaSyDpnGR_nxQ3hQyrnrvNHsmOPkBEYaM6xmg'

place_id = 'ChIJyYidmw1ZIkcRvels1i3CmzU'

def fetch_reviews(api_key, place_id):
    reviews = []
    url = f'https://maps.googleapis.com/maps/api/place/details/json?place_id={place_id}&fields=reviews&key={api_key}'

    response = requests.get(url)
    data = response.json()

    # Extract reviews from the response
    if 'result' in data:
        for review in data['result'].get('reviews', []):
            # Convert Unix timestamp to readable format
            readable_time = pd.to_datetime(review['time'], unit='s')
            reviews.append({
                'author_name': review['author_name'],
                'rating': review['rating'],
                'text': review['text'],
                'time': readable_time
            })

    return reviews

# Fetch the reviews
reviews = fetch_reviews(api_key, place_id)

# Convert reviews to DataFrame
reviews_df = pd.DataFrame(reviews)

print(reviews_df)

reviews_df.to_csv('customer_reviews.csv', index=False)
