==> 05_SQL_integration.ipynb <==
# Analysis with SQL

import sqlite3
import pandas as pd

conn = sqlite3.connect('pizzeria.db')
cursor = conn.cursor()

sales_df = pd.read_csv('sales_transactions.csv')
menu_df = pd.read_csv('menu_data.csv')
inventory_df = pd.read_csv('inventory_ingredients.csv')
reviews_df = pd.read_csv('customer_reviews.csv')

sales_df.to_sql('sales', conn, if_exists='replace', index=False)
menu_df.to_sql('menu', conn, if_exists='replace', index=False)
inventory_df.to_sql('inventory', conn, if_exists='replace', index=False)
reviews_df.to_sql('reviews', conn, if_exists='replace', index=False)

conn.commit()
conn.close()

### Total Sales by Item (Revenue per Menu Item)

conn = sqlite3.connect('pizzeria.db')
query = '''
SELECT m.Item_Name, SUM(s.Quantity * s.Price) AS Total_Revenue
FROM sales s
JOIN menu m ON s.Item_ID = m.Item_ID
GROUP BY m.Item_Name
ORDER BY Total_Revenue DESC;
'''
sales_by_item = pd.read_sql(query, conn)
sales_by_item.to_csv('sales_by_item.csv', index=False)  # index=False to avoid writing row numbers
conn.close()

# Show the result
print(sales_by_item)

### Total Monthly Sales

conn = sqlite3.connect('pizzeria.db')
query = '''
SELECT strftime('%Y-%m', Date) AS Month, SUM(Quantity * Price) AS Total_Revenue
FROM sales
GROUP BY Month
ORDER BY Month;
'''
monthly_sales = pd.read_sql(query, conn)
monthly_sales.to_csv('monthly_sales.csv', index=False)  # index=False to avoid writing row numbers

conn.close()

# Show the result
print(monthly_sales)

### Most Used Ingredients

conn = sqlite3.connect('pizzeria.db')
query = '''
SELECT i.Ingredient, SUM(s.Quantity) AS Total_Usage
FROM sales s
JOIN inventory i ON s.Item_ID = i.Item_ID
GROUP BY i.Ingredient
ORDER BY Total_Usage DESC;
'''
inventory_usage = pd.read_sql(query, conn)
inventory_usage.to_csv('inventory_usage.csv', index=False)  # index=False to avoid writing row numbers

conn.close()

# Show the result
print(inventory_usage)

### Review Rating Distribution

conn = sqlite3.connect('pizzeria.db')
query = '''
SELECT rating, COUNT(rating) AS Rating_Count
FROM reviews
GROUP BY rating
ORDER BY rating DESC;
'''
rating_distribution = pd.read_sql(query, conn)
conn.close()

# Show the result
print(rating_distribution)
